Update Battlefield4Seeder.au3
Per Avatar:

"This version is what the seeders are currently running. The [previous] version on Github is newer, but in a half-finished state and wouldn't be a good starting point for someone who's new to the language. There isn't anything really new in the [previous] version on github so it can be safely ignored and replaced by [this version]."